Make mana circles and limit breaks cost

+++ DragaliaAPI.Integration.Test/Dragalia/CharaTest.cs
@@ -119,7 +119,7 @@ public class CharaTest : TestFixture
                 new CharaBuildupManaRequest(
                     Charas.Celliera,
                     new List<int>() { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 },
-                    (int)CharaUpgradeMaterialTypes.Standard
+                    CharaUpgradeMaterialTypes.Standard
                 )
             )
         ).data;
